Revert "Open/close on each write"
This reverts commit df0649fb8f5b80fb76e1151da17a64115a6710ee.

+++ JM/CF/Scripts/4_World/CommunityFramework/ModStorage/CF_ModStorageModule.c
@@ -72,10 +72,15 @@ class CF_ModStorageModule : CF_ModuleWorld
 	{
 		super.OnMissionFinish(sender, args);
 
-		if (m_LockFileHandle)
+		if (m_Serializer)
 		{
-			CloseFile(m_LockFileHandle);
-			DeleteFile(m_LockFilePath);
+			m_Serializer.Close();
+
+			if (m_LockFileHandle)
+			{
+				CloseFile(m_LockFileHandle);
+				DeleteFile(m_LockFilePath);
+			}
 		}
 	}
 
@@ -110,15 +115,8 @@ class CF_ModStorageModule : CF_ModuleWorld
 
 		CF_Log.Debug("  added");
 
-		if (!m_Serializer.Open(m_FilePath, FileMode.APPEND))
-		{
-			Error("Failed to open \"" + m_FilePath + "\" for appending!");
-		}
-		else
-		{
-			m_Serializer.Write(id);
-			m_Serializer.Close();
-		}
+		m_Serializer.Write(id);
+		_Flush();
 	}
 
 	/**
@@ -191,7 +189,9 @@ class CF_ModStorageModule : CF_ModuleWorld
 				CF_Log.Error("Could not create lockfile \"%1\"!", m_LockFilePath);
 		}
 
-		if (!m_Serializer)
+		if (m_Serializer)
+			m_Serializer.Close();
+		else
 			m_Serializer = new FileSerializer();
 
 		// Clear existing ids
@@ -251,14 +251,14 @@ class CF_ModStorageModule : CF_ModuleWorld
 
 				CopyFile(m_FilePath, backup1);
 			}
+			
+			if (!m_Serializer.Open(m_FilePath, FileMode.APPEND))
+				_CriticalError("Failed to open \"" + m_FilePath + "\" for appending!");
 		}
 		else
 		{
-			//! Create the file
 			if (!m_Serializer.Open(m_FilePath, FileMode.WRITE))
 				_CriticalError("Failed to open \"" + m_FilePath + "\" for writing!");
-			else
-				m_Serializer.Close();
 		}
 	}
 
@@ -266,6 +266,13 @@ class CF_ModStorageModule : CF_ModuleWorld
 	{
 		Error("[CRITICAL]\t" + message);
 		GetGame().RequestExit(1);
+	}
+
+	private void _Flush()
+	{
+		m_Serializer.Close();
+		if (!m_Serializer.Open(m_FilePath, FileMode.APPEND))
+			Error("Failed to open \"" + m_FilePath + "\" for appending!");
 	}
 
 	/**
